+++ auralogger/__init__.py
@@ -11,7 +11,7 @@ from auralogger.cli.log_styles import (
     resolve_log_style_spec,
     style_map_from_config_entries,
 )
-from auralogger.server.aura_log import auralogger, aura_log, close_aura_log_socket
+from auralogger.server.aura_log import Auralogger, aura_log, close_aura_log_socket
 from auralogger.server.proj_auth import fetch_proj_auth_config, fetch_proj_auth_payload
 from auralogger.utils.backend_origin import build_proj_auth_url, build_project_logs_url
 from auralogger.utils.http_utils import parse_error_body
@@ -20,7 +20,7 @@ from auralogger.utils.parser import ParsedFilter, ParsedGetLogsCommand, parse_co
 __all__ = [
     "aura_log",
     "close_aura_log_socket",
-    "auralogger",
+    "Auralogger",
     "run_get_logs",
     "normalize_and_validate_filters",
     "ApiLogFilter",